+++ branches/avidemux_2.5_branch_gruntster/avidemux/ADM_libraries/ADM_ffmpeg/ADM_lavcodec/i386/fft_sse.c
@@ -48,8 +48,8 @@ static void print_v4sf(const char *str, __m128 a)
 void ff_fft_calc_sse(FFTContext *s, FFTComplex *z)
 {
     int ln = s->nbits;
-    long i, j;
-    long nblocks, nloops;
+    intptr_t i, j;
+    intptr_t nblocks, nloops;
     FFTComplex *p, *cptr;
 
     asm volatile(
@@ -142,7 +142,7 @@ void ff_fft_calc_sse(FFTContext *s, FFTComplex *z)
 void ff_imdct_calc_sse(MDCTContext *s, FFTSample *output,
                        const FFTSample *input, FFTSample *tmp)
 {
-    long k, n8, n4, n2, n;
+    intptr_t k, n8, n4, n2, n;
     const uint16_t *revtab = s->fft.revtab;
     const FFTSample *tcos = s->tcos;
     const FFTSample *tsin = s->tsin;
